"Use react-navigation to switch auth mode in AuthContent component"
This commit updates the AuthContent component to use the useNavigation hook from @react-navigation/native to navigate between the login and signup screens based on the current auth mode. Previously, the function to switch auth modes was commented out with a TODO. The code also ensures that the text for the FlatButton is updated based on the current auth mode.